pi_server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.2
aiosqlite==0.20.0
paho-mqtt==2.1.0
google-generativeai==0.8.3

